+++ data/products.js
@@ -101,22 +101,6 @@ export function recommendProduct(id, username) {
   })
 }
 
-export function likeProduct(productId) {
-  return fetchWithoutResponse(`products/${productId}/like`, {
-    method: 'POST',
-    headers: {
-      Authorization: `Token ${localStorage.getItem('token')}`,
-      'Content-Type': 'application/json'
-    },
-  })
-}
 
-export function unLikeProduct(productId) {
-  return fetchWithoutResponse(`products/${productId}/unlike`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Token ${localStorage.getItem('token')}`,
-      'Content-Type': 'application/json'
-    },
-  })
-}
+
+
